backend/requirements.txt
# FastAPI 后端依赖包
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP 客户端
httpx==0.25.2
requests==2.31.0

# 工具库
python-dotenv==1.0.0
email-validator==2.1.0
redis==5.0.1

# AI 相关（可选）
openai==1.3.7
langchain==0.0.350
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# 文件处理
python-magic==0.4.27
PyPDF2==3.0.1
docx2txt==0.8

# WebSocket
websockets==12.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
